Relocate cache-related classes to mcpe\cache namespace
